Hopefully fix the tests
