+++ tests/frontend/test_headers.py
@@ -1,7 +1,6 @@
-import pytest
+from utils.header_assertions import *
+
 import requests
-import sh
-import test_headers
 
 
 def test_server_header_not_returned(domain):
@@ -10,8 +9,8 @@ def test_server_header_not_returned(domain):
     Other than that it doesn't bring anything useful. Therefore we should hide it.  
     """
 
-    res = request.get(domain)
-    test_headers.assert_no_server_header(res)
+    res = requests.get(domain)
+    assert_no_server_header(res)
 
 
 def test_ip_resolver_not_returned(domain):
@@ -21,8 +20,8 @@ def test_ip_resolver_not_returned(domain):
     away the fact that we are using nginx.
     """
 
-    res = request.get(domain)
-    test_headers.assert_no_x_resolver_ip_header(res)
+    res = requests.get(domain)
+    assert_no_x_resolver_ip_header(res)
 
 
 def test_hsts_present(domain):
@@ -32,8 +31,8 @@ def test_hsts_present(domain):
     run on the http protocol)
     """
 
-    res = request.get(domain)
-    test_headers.assert_hsts_present(res)
+    res = requests.get(domain)
+    assert_hsts_present(res)
 
 
 def test_csp_present(domain):
@@ -42,8 +41,8 @@ def test_csp_present(domain):
     and configured.
     """
 
-    res = request.get(domain)
-    test_headers.assert_csp_present(res)
+    res = requests.get(domain)
+    assert_csp_present(res)
 
 
 def test_click_jacking_protection(domain):
@@ -53,8 +52,8 @@ def test_click_jacking_protection(domain):
     our page in an iframe, frame or object.
     """
 
-    res = request.get(domain)
-    test_headers.assert_click_jacking_protection(res)
+    res = requests.get(domain)
+    assert_click_jacking_protection(res)
 
 
 def test_XSS_protection_present(domain):
@@ -63,8 +62,8 @@ def test_XSS_protection_present(domain):
     the problems solved by the X-XSS-Protection header, but it doesn't hurt to have it as well
     """
 
-    res = request.get(domain)
-    test_headers.assert_XSS_protection_present(res)
+    res = requests.get(domain)
+    assert_XSS_protection_present(res)
 
 
 def test_content_type_options_present(domain):
@@ -76,9 +75,8 @@ def test_content_type_options_present(domain):
     the web server, assuming the web server accepts JPGs. Consequently, the browser will render it as
     an HTML file therefore providing the attacker with the possibility to execute XSS.
     """
-    res = request.get(domain)
-    test_headers.assert_content_type_options_present(
-        res)
+    res = requests.get(domain)
+    assert_content_type_options_present(res)
 
 
 def test_referrer_policy_present(domain):
@@ -87,8 +85,8 @@ def test_referrer_policy_present(domain):
     makes sure we never receive a downgraded (i.e. http) link
     """
 
-    res = request.get(domain)
-    test_headers.assert_referrer_policy_present(res)
+    res = requests.get(domain)
+    assert_referrer_policy_present(res)
 
 
 def test_feature_policy_present(domain):
@@ -97,8 +95,8 @@ def test_feature_policy_present(domain):
     interest of better security and privacy
     """
 
-    res = request.get(domain)
-    test_headers.assert_feature_policy_present(res)
+    res = requests.get(domain)
+    assert_feature_policy_present(res)
 
 
 def test_expect_ct_present(domain):
@@ -106,5 +104,5 @@ def test_expect_ct_present(domain):
     Certificate Transparency
     """
 
-    res = request.get(domain)
-    test_headers.assert_expect_ct_present(res)
+    res = requests.get(domain)
+    assert_expect_ct_present(res)
